examples/figma2code/requirements.txt
vllm==0.11.0
rich==14.1.0
transformers==4.57.1
huggingface-hub==0.34.4
gradio==5.43.1
matplotlib==3.10.7
datasets==4.4.1
playwright==1.55.0
beautifulsoup4==4.13.4
CairoSVG==2.8.2
scikit-image==0.25.2
selenium==4.34.2
webdriver-manager==4.0.2
ipython==9.10.0
lpips==0.1.4
openai-clip==1.0.1
lxml==6.0.2

openai==2.7.1
pandas==2.3.1
python-dotenv==1.1.1
requests==2.32.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
scipy==1.16.3
tqdm==4.67.1



